move message_event functionality to message_events.py
